- change dbName

+++ src/views/menuView.js
@@ -2,6 +2,7 @@ const { Menu } = require("../models/Menu");
 const { Article } = require("../models/Article");
 const { mongoose, isValidObjectId } = require("mongoose");
 const { Product } = require("../models/Product");
+const { Restaurant } = require("../models/Restaurant");
 
 const errors = {
     invalidId: (() => {
@@ -17,6 +18,11 @@ const errors = {
     idNotFound: (() => {
         const err = Error("Id not found");
         err.statusCode = 404;
+        return err;
+    })(),
+    Unauthorized: (() => {
+        const err = Error("Access denied");
+        err.statusCode = 403;
         return err;
     })(),
   }
@@ -67,6 +73,7 @@ module.exports = {
 
         let elements = [];
         const articles = await Article.find({restaurantId: restaurantId});
+        if(articles.length <= 0) error.idNotFound;
         for(let i = 0; i < articles.length; i++) {
             const menu = await Menu.findOne({articleId: articles[i].id});
             if(menu) {
@@ -80,16 +87,16 @@ module.exports = {
                     imageUrl: articles[i].imageUrl, 
                     productIdList: menu.productIdList
                 })
-                return elements;
             }
-            else return errors.idNotFound;
         }
+        return elements;
     },
     getMenuByName: async(req, res) => {
         const { menuName } = req.params;
 
         let elements = [];
         const articles = await Article.find({name: { '$regex' : menuName, '$options' : 'i' }});
+        if(articles.length <= 0) return error.idNotFound;
         for(let i = 0; i < articles.length; i++) {
             const menu = await Menu.findOne({articleId: articles[i].id});
             if(menu) {
@@ -103,16 +110,17 @@ module.exports = {
                     imageUrl: articles[i].imageUrl, 
                     productIdList: menu.productIdList
                 })
-                return elements;
             }
-            else return errors.idNotFound;
         }
+        return elements;
     },
     createMenu: async(req, res) => {
         const { name, price, description, restaurantId, imageUrl, productIdList } = req.body;
         if (!isValidObjectId(restaurantId)) return errors.invalidId;
+        const restaurant = await Restaurant.findById(restaurantId);
+        if(req.query.roleLabel != 'admin' && req.query.userId != restaurant.restaurantOwnerId) return errors.Unauthorized;
 
-        // check if productId exist and convert it in ObjectId
+        // check if productId exist
         for(let i = 0; i < productIdList.length; i++) {
             if (!isValidObjectId(productIdList[i])) return errors.invalidId;
             const product = await Product.exists({_id: productIdList[i]});
@@ -138,18 +146,23 @@ module.exports = {
         const { menuId } = req.params;
         if (!isValidObjectId(menuId)) return errors.invalidId;
         
-        const menu = await Menu.findByIdAndDelete(menuId);
-        if(menu) {
-            await Article.findByIdAndDelete(menu.articleId);
-        }
-        else return errors.idNotFound;
+        const menu = await Menu.findById(menuId);
+        const article = await Article.findById(menu.articleId);
+        const restaurant = await Restaurant.findById(article.restaurantId);
+        if(req.query.roleLabel != 'admin' && req.query.userId != restaurant.restaurantOwnerId) return errors.Unauthorized;
+
+        if(!menu) return errors.idNotFound;
+        await Menu.findByIdAndDelete(menuId);
+        await Article.findByIdAndDelete(menu.articleId);
         return `menu ${menuId} deleted`;
     },
     updateMenu: async(req, res) => {
         const { menuId } = req.params;
         const { name, price, description, restaurantId, imageUrl, productIdList } = req.body;
         if (!isValidObjectId(menuId)) return errors.invalidId;
-        
+        const restaurant = await Restaurant.findById(restaurantId);
+        if(req.query.roleLabel != 'admin' && req.query.userId != restaurant.restaurantOwnerId) return errors.Unauthorized;
+
         // check if productId exist and convert it in ObjectId
         for(let i = 0; i < productIdList.length; i++) {
             if (!isValidObjectId(productIdList[i])) return errors.invalidId;
@@ -188,14 +201,17 @@ module.exports = {
             const product = await Product.exists({_id: productIdList[i]});
             if(!product) return `${productIdList[i]} doesn't exist`; 
         }
+        const menu = await Menu.findById(menuId);
+        const article = await Article.findById(menu.articleId);
+        const restaurant = await Restaurant.findById(article.restaurantId);
+        if(req.query.roleLabel != 'admin' && req.query.userId != restaurant.restaurantOwnerId) return errors.Unauthorized;
 
-        const menu = await Menu.findOneAndUpdate({
+        if(!menu) return errors.idNotFound;
+        await Menu.findOneAndUpdate({
             _id: menuId
         }, {
             productIdList: productIdList
         });
-
-        if(menu) return `productIdList in menu ${menuId} updated`;
-        else return errors.idNotFound;
+        return `productIdList in menu ${menuId} updated`;
     }
 }
